refactor: Break up dynamics module
This splits the dynamics module into smaller sections which make it more
manageable to edit and organise.
